refactor: deprecate public `settings` module and its symbols
